+++ banking_system/src/main/java/edu/sjsu/cmpe202/banking_system/account_creation/saving_account/SavingAccountRepository.java
@@ -3,6 +3,6 @@ package edu.sjsu.cmpe202.banking_system.account_creation.saving_account;
 import edu.sjsu.cmpe202.banking_system.account_creation.checking_account.CheckingAccount;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SavingAccountRepository extends JpaRepository<SavingAccount, Integer> {
+public interface SavingAccountRepository extends JpaRepository<SavingAccount, Long> {
      SavingAccount findByUserId (int user_id);
 }
